refactor: centralize battle orientation setup
