+++ src/pages/battleJudoka.html
@@ -152,23 +152,11 @@
           </ul>
         </nav>
       </footer>
-      <script type="module" src="../helpers/setupBattleOrientation.js"></script>
-      <script type="module" src="../helpers/setupBottomNavbar.js"></script>
-      <script type="module" src="../helpers/setupDisplaySettings.js"></script>
-
       <noscript>
         <p class="noscript-warning">
           JU-DO-KON! requires JavaScript to run. Please enable JavaScript to play.
         </p>
       </noscript>
-
-      <script nomodule>
-        alert(
-          "Your browser does not support modern JavaScript. Please update your browser to play JU-DO-KON!"
-        );
-      </script>
-      <script type="module" src="../helpers/setupSvgFallback.js"></script>
-      <script type="module" src="../helpers/setupClassicBattleHomeLink.js"></script>
       <script type="module" src="../helpers/classicBattlePage.js"></script>
     </div>
     <div id="snackbar-container" role="status" aria-live="polite"></div>
